refactor: make pretty hooks for consuming the okr stuff
